Added mImportAnim.ImportAnimOperator to the module.
Implements a Shadow Animation CSV importer. Create curves quickly and
apply them to our premade Shadow armature. This way, the user can use
our live preview and take animation on the same armature in the same
scene.

Also, it allows the user to setup a retargeting system one time and them
load multiple takes. The takes are loaded as individual armature
"actions" and can be selected in the Blender Animation Action Editor.